scripts/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.11.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
